sync_dhis2_to_dhis2/requirements.txt
openhexa.sdk>=1.0.0
openhexa-toolbox[dhis2]>=0.1.0